Scene

+++ Engine/src/Resource/Loader/GLTFLoader.h
@@ -50,9 +50,9 @@ public:
             size_t uv = std::hash<float>()(k.texCoord.x) ^
                 (std::hash<float>()(k.texCoord.y) << 1);
 
-            return ((std::hash<float>()(pos)
-                ^ (std::hash<float>()(nor) << 1)) >> 1)
-                ^ (std::hash<float>()(uv) << 1);
+            return ((std::hash<size_t>()(pos)
+                ^ (std::hash<size_t>()(nor) << 1)) >> 1)
+                ^ (std::hash<size_t>()(uv) << 1);
         }
     };
