+++ Engine/src/Resource/Loader/OBJLoader.cpp
@@ -302,7 +302,7 @@ void OBJLoader::ComputeVertices(Mesh& mesh)
         }
         else
         {
-            const uint32_t id = mesh.finalVertices.size();
+            const uint32_t id = static_cast<uint32_t>(mesh.finalVertices.size());
             mesh.finalVertices.push_back(v);
             mesh.finalIndices.push_back(id);
             hashed_vertices[v] = id;
